feat: introduce zero-dependency sqlite_move and standardize cross-platform toolchains
Add zero-dependency `sqlite_move` to eliminate `<utility>` dependencies, qualify pseudo-destructors to silence `-Wdelete-non-abstract-non-virtual-dtor`, and standardize Clang64/AddressSanitizer and MSVC `/W4` toolchain targets across all Makefiles and batch runners.

Key architectural additions:
- include/sqlite3_allocator.hpp:
  * Implemented `sqlite_move` rvalue reference cast template to replace `std::move` across all freestanding `-nostdlib++` components.
  * Included `<stddef.h>` for standard, portable `size_t` definitions without standard library runtime dependencies.
  * Qualified explicit destructor invocations (`ptr->T::~T()`) in `sqlite_delete`, `sqlite_destroy_at`, and `sqlite_destroy_n` to silence non-virtual destructor warnings.
- include/sqlite3_aggregate.hpp:
  * Suppressed and documented MSVC warning C4324 around `SqliteAggregateDetail::AggregateHolder<T>` explaining intentional `alignas(T)` storage padding.
  * Qualified destructor invocation in `final_proxy` (`holder->instance()->T::~T()`).
- include/sqlite3_value.hpp:
  * Included `sqlite3_allocator.hpp` to ensure `sqlite_move` and memory helpers are available downstream.

Build system & Toolchain standardizations:
- Makefile & tests/*/Makefile:
  * Standardized unified OS detection and toolchain resolution across root and all 20 subfolder Makefiles.
  * Configured automated detection and invocation of `/clang64/bin/clang++` and `/clang64/bin/clang` on MSYS2/Windows, ensuring dynamic AddressSanitizer runtime (`libclang_rt.asan_dynamic-x86_64.dll`) resolution.
  * Expanded root `clean` target to clean all 19 test subdirectories and examples.
- make.bat & tests/*/make.bat:
  * Enforced MSVC `/W4` warning levels, standard library prohibition (`/NODEFAULTLIB:msvcprt.lib`), and clean targets across batch test runners.

Testing & Documentation:
- tests/cpp_buffer/test_buffer.cpp, tests/cpp_db/test_db.cpp, tests/cpp_statement/test_statement.cpp, tests/cpp_blob_stream/test_blob_stream.cpp:
  * Replaced all occurrences of `std::move` and `static_cast<...&&>` with `sqlite_move`.
  * Silenced Clang `-Wself-move` diagnostic pragmas during intentional self-assignment tests.
- tests/ext_state/c_extension/myext.c & tests/ext_state/cpp_extension/myext.cpp:
  * Added formal parameter discards to silence MSVC unreferenced parameter warnings (C4100).
- tests/cpp_extension/test_ext_stateful.cpp:
  * Added explicit type cast on string length to silence MSVC C4267 conversion warning.
- SETUP.md:
  * Created complete multi-platform developer setup guide covering Linux (GCC), macOS (Apple Clang), Windows MSYS2 (Clang64 / AddressSanitizer), and Windows Native MSVC (`cl.exe` / `make.bat`).
- README.md, ARCHITECTURE.md, docs/ALLOCATOR_ARCHITECTURE.md, docs/ALLOCATOR_README.md:
  * Updated allocator and architecture documentation to detail `sqlite_move`, zero-dependency type traits, `SQLITE_FAST_MEMCPY`, and dual build system execution.

+++ tests/cpp_extension/test_ext_stateful.cpp
@@ -93,7 +93,7 @@ struct StatefulTaggedConcat : public SqliteAggregateBase<void> {
             res.appendall(lock->session_tag);
             res.appendall(":");
         }
-        res.append(buffer, len);
+        res.append(buffer, static_cast<int>(len));
         res.result(ctx);
     }
 };
